requirements.txt
prometheus-client==0.5.0
python-dateutil==2.7.5
six==1.12.0
requests==2.21.0
urllib3==1.24.1

